Adding Basic Sorted Set Operations (zadd, zrem, zrange)
The Sorted Set is stored as a LinkedHashMap, with the map's key being the value in the set, and the map's value being the score. This ensures that the values inserted into the set are unique (map's key is unique), and we can easily update the score for a given value. Caveat: every time we insert a new value (or update a value's score), we have to re-sort the LinkedHashMap.