sscanIterates test should not count iterations
According to the Redis documentation (https://redis.io/commands/scan), there is no guarantee on the count of items returned by scan, sscan or other related methods. Even specifying a COUNT, the scan methods only use this as a suggestion.

As a result, the sscanIterates test that was counting the number of times the sscan iterates would sometimes fail (if sscan returned more than 10 elements at a time, which is completely valid). Instead of counting the number of iterations, which can never be predicted, I've modified the test to instead make sure all elements were returned after the iterations complete.

Also removing the `jedis.del(key)` line as the entire DB is erased at the start of the test already.